// packages/web/src/api/models/domain.ts
export interface UserDomain {
    userId?: number;
    twitterId?: string;
    name?: string;
    username?: string;
    profileImageUrl?: string;
}

export interface WhitelistDomain {
    user: UserDomain;
    referralCode?: string;
    status: string;
}

export interface ReferralDomain {
    referrerId: number;
    referrerTwitterId: string;
    refereeId: number;
    refereeTwitterId: string;
}